+++ HerraduraCli/primitives.py
@@ -30,6 +30,8 @@ hfscx_256_ds           = _s.hfscx_256_ds
 hmac_hfscx_256         = _s.hmac_hfscx_256
 hske_nl_aead_encrypt   = _s.hske_nl_aead_encrypt
 hske_nl_aead_decrypt   = _s.hske_nl_aead_decrypt
+hske_nl_v2_duplex_encrypt = _s.hske_nl_v2_duplex_encrypt
+hske_nl_v2_duplex_decrypt = _s.hske_nl_v2_duplex_decrypt
 _HFSCX256_IV_BYTES     = _s._HFSCX256_IV_BYTES
 _RNL_KDF_DC_256        = _s._RNL_KDF_DC_256